O3-1622: Remove Cucumber from the RefApp 3.x Testing tooling
