--- packages/cypress/cypress/integration/cucumber/step_definitions/refapp-3.x/01-login/login.js
@@ -1,35 +1,0 @@
-import {Given, Then, When} from '@badeball/cypress-cucumber-preprocessor';
-
-Given('user arrives at the login page', () => {
-    cy.visit('/login');
-})
-
-When('the user logs in with {string} and {string} to the {string}', (username, password, location) => {
-    cy.getByLabel('Username').type(username)
-    cy.contains('Continue').click({force: true});
-    cy.getByLabel('Password').type(password)
-    cy.contains('Log in').click({force: true});
-    // The user will be redirected to the location selection page if the credentials are correct.
-    // The following artificial wait is to wait until it changes the url.
-    cy.wait(2000);
-    // If the url includes "/location" that means the user is on the location selection page. If so, pick the location.
-    cy.url().then(($url) => {
-        if($url.includes("/location")) {
-            cy.contains(location).click({force: true});
-            cy.contains('Confirm').click({force: true});
-        }
-    })
-})
-
-Then('the user should be {string} to login', (ability) => {
-    switch (ability) {
-        case "able":
-            cy.url().should('include', '/home');
-            break;
-        case "unable":
-            cy.contains("Invalid username or password");
-            break;
-        default:
-            throw new Error(`Ability '${ability}' is not supported`);
-    }
-});
