--- packages/cypress/cypress/integration/cucumber/step_definitions/refapp-3.x/06-patient-search/patient-search.js
@@ -1,17 +1,0 @@
-import {Then, When} from '@badeball/cypress-cucumber-preprocessor';
-
-When('John Doe is registered on the system', () => {
-    cy.createPatient();
-});
-
-When('the user search for {string}', patientName => {
-    cy.get('button[name=SearchPatientIcon]').click();
-    cy.getByPlaceholder('Search for a patient by name or identifier number').type(patientName);
-    // Adding an artificial wait because typed texts takes some time to get updated
-    cy.wait(1000);
-    cy.getByPlaceholder('Search for a patient by name or identifier number').type('{enter}');
-})
-
-Then('the result should be {string}', result => {
-    cy.contains(result);
-});
